requirements.txt
streamlit==1.36.0
pandas==2.2.2
scikit-learn==1.5.0
PyMuPDF==1.24.7
python-docx==1.1.2
spacy==3.7.5
nltk==3.8.1
matplotlib==3.8.4
seaborn==0.13.2
python-json-logger==2.0.7